modular i3

+++ i3/conf/build.sh
@@ -1,3 +1,13 @@
 #!/bin/bash
-rm ~/.dotfiles/i3/config
-cat ~/.dotfiles/i3/conf/client/$(hostname)/head ~/.dotfiles/i3/conf/common/base > ~/.dotfiles/i3/config
+COMMON=~/.dotfiles/i3/conf/common/
+CLIENT=~/.dotfiles/i3/conf/client/$(hostname)/
+
+if [ -f "~/.dotfiles/i3/config" ]; then
+    rm ~/.dotfiles/i3/config
+fi
+
+if [ ! -d $CLIENT ]; then
+    cp -r ~/.dotfiles/i3/conf/client/default ~/.dotfiles/i3/conf/client/$(hostname)
+fi
+
+cat ${CLIENT}header ${COMMON}base ${CLIENT}assignments ${COMMON}footer > ~/.dotfiles/i3/config
